pyproject.toml
[tool.poetry]
name = "course-5"
version = "0.1.0"
description = ""
authors = ["Aleksey Belyaev <lesha.belyaev.73@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
